test: extract batch processing test
