+++ engine/src/test/java/io/camunda/zeebe/streamprocessor/StreamProcessorTest.java
@@ -231,74 +231,6 @@ public final class StreamProcessorTest {
   }
 
   @Test
-  public void shouldProcessFollowUpEventsAndCommands() {
-    // given
-    final var defaultRecordProcessor = streamPlatform.getDefaultMockedRecordProcessor();
-    final var resultBuilderCaptor = ArgumentCaptor.forClass(ProcessingResultBuilder.class);
-    when(defaultRecordProcessor.process(any(), resultBuilderCaptor.capture()))
-        .thenAnswer(
-            (invocation) -> {
-              final var resultBuilder = resultBuilderCaptor.getValue();
-              resultBuilder.appendRecordReturnEither(
-                  1,
-                  RecordType.EVENT,
-                  ACTIVATE_ELEMENT,
-                  RejectionType.NULL_VAL,
-                  "",
-                  Records.processInstance(1));
-              resultBuilder.appendRecordReturnEither(
-                  2,
-                  RecordType.COMMAND,
-                  ACTIVATE_ELEMENT,
-                  RejectionType.NULL_VAL,
-                  "",
-                  Records.processInstance(1));
-              return resultBuilder.build();
-            })
-        .thenAnswer(
-            (invocation) -> {
-              final var resultBuilder = resultBuilderCaptor.getValue();
-              resultBuilder.appendRecordReturnEither(
-                  3,
-                  RecordType.EVENT,
-                  ACTIVATE_ELEMENT,
-                  RejectionType.NULL_VAL,
-                  "",
-                  Records.processInstance(1));
-              return resultBuilder.build();
-            });
-
-    streamPlatform.startStreamProcessor();
-
-    // when
-    streamPlatform.writeBatch(command().processInstance(ACTIVATE_ELEMENT, RECORD));
-
-    // then
-    verify(defaultRecordProcessor, TIMEOUT.times(2)).process(any(), any());
-    await("Last written position should be updated")
-        .untilAsserted(
-            () -> assertThat(streamPlatform.getLogStream().getLastWrittenPosition()).isEqualTo(4));
-    await("Last processed position should be updated")
-        .untilAsserted(
-            () ->
-                assertThat(
-                        streamPlatform.getStreamProcessor().getLastProcessedPositionAsync().join())
-                    .isEqualTo(1));
-
-    final var logStreamReader = streamPlatform.getLogStream().newLogStreamReader();
-    logStreamReader.seekToFirstEvent();
-    final var firstRecord = logStreamReader.next();
-    assertThat(firstRecord.getSourceEventPosition()).isEqualTo(-1);
-    final var firstRecordPosition = firstRecord.getPosition();
-
-    await("should write follow up events")
-        .untilAsserted(() -> assertThat(logStreamReader.hasNext()).isTrue());
-    while (logStreamReader.hasNext()) {
-      assertThat(logStreamReader.next().getSourceEventPosition()).isEqualTo(firstRecordPosition);
-    }
-  }
-
-  @Test
   public void shouldSetSourcePointerForFollowUpRecords() {
     // given
     final var defaultRecordProcessor = streamPlatform.getDefaultMockedRecordProcessor();
